Refactored the GA event sender into its own module in lib. Added test analytics for openurl, with some other improvements
