+++ portality/view/api_v1.py
@@ -7,7 +7,8 @@ from portality.api.v1 import DiscoveryApi, DiscoveryException
 from portality.api.v1 import ApplicationsCrudApi, ArticlesCrudApi, JournalsCrudApi, ApplicationsBulkApi, ArticlesBulkApi
 from portality.api.v1 import jsonify_models, jsonify_data_object, Api400Error, Api401Error, Api404Error, created, no_content, bulk_created
 from portality.core import app
-from portality.decorators import api_key_required, api_key_optional, swag, track_analytics
+from portality.decorators import api_key_required, api_key_optional, swag
+from portality.lib.analytics import google_analytics_event
 
 import json
 
@@ -55,7 +56,7 @@ def docs():
 @swag(swag_summary='Search your applications <span class="red">[Authenticated, not public]</span>', swag_spec=DiscoveryApi.get_application_swag())  # must be applied after @api_key_(optional|required) decorators. They don't preserve func attributes.
 @blueprint.route("/search/applications/<path:search_query>")
 @api_key_required
-@track_analytics('API Hit', 'Search applications', record_value_of_which_arg='search_query')
+@google_analytics_event('API Hit', 'Search applications', record_value_of_which_arg='search_query')
 def search_applications(search_query):
     # get the values for the 2 other bits of search info: the page number and the page size
     page = request.values.get("page", 1)
@@ -84,7 +85,7 @@ def search_applications(search_query):
 
 @swag(swag_summary='Search journals', swag_spec=DiscoveryApi.get_journal_swag())  # must be applied after @api_key_(optional|required) decorators. They don't preserve func attributes.
 @blueprint.route('/search/journals/<path:search_query>')
-@track_analytics('API Hit', 'Search journals', record_value_of_which_arg='search_query')
+@google_analytics_event('API Hit', 'Search journals', record_value_of_which_arg='search_query')
 def search_journals(search_query):
     # get the values for the 2 other bits of search info: the page number and the page size
     page = request.values.get("page", 1)
@@ -113,7 +114,7 @@ def search_journals(search_query):
 
 @swag(swag_summary='Search articles', swag_spec=DiscoveryApi.get_article_swag())  # must be applied after @api_key_(optional|required) decorators. They don't preserve func attributes.
 @blueprint.route('/search/articles/<path:search_query>')
-@track_analytics('API Hit', 'Search articles', record_value_of_which_arg='search_query')
+@google_analytics_event('API Hit', 'Search articles', record_value_of_which_arg='search_query')
 def search_articles(search_query):
     # get the values for the 2 other bits of search info: the page number and the page size
     page = request.values.get("page", 1)
@@ -147,7 +148,7 @@ def search_articles(search_query):
 @blueprint.route("/applications", methods=["POST"])
 @api_key_required
 @swag(swag_summary='Create an application <span class="red">[Authenticated, not public]</span>', swag_spec=ApplicationsCrudApi.create_swag())  # must be applied after @api_key_(optional|required) decorators. They don't preserve func attributes.
-@track_analytics('API Hit', 'Create application')
+@google_analytics_event('API Hit', 'Create application')
 def create_application():
     # get the data from the request
     try:
@@ -164,7 +165,7 @@ def create_application():
 @blueprint.route("/applications/<application_id>", methods=["GET"])
 @api_key_required
 @swag(swag_summary='Retrieve an application <span class="red">[Authenticated, not public]</span>', swag_spec=ApplicationsCrudApi.retrieve_swag())  # must be applied after @api_key_(optional|required) decorators. They don't preserve func attributes.
-@track_analytics('API Hit', 'Retrieve application', record_value_of_which_arg='application_id')
+@google_analytics_event('API Hit', 'Retrieve application', record_value_of_which_arg='application_id')
 def retrieve_application(application_id):
     a = ApplicationsCrudApi.retrieve(application_id, current_user)
     return jsonify_models(a)
@@ -172,7 +173,7 @@ def retrieve_application(application_id):
 @blueprint.route("/applications/<application_id>", methods=["PUT"])
 @api_key_required
 @swag(swag_summary='Update an application <span class="red">[Authenticated, not public]</span>', swag_spec=ApplicationsCrudApi.update_swag())  # must be applied after @api_key_(optional|required) decorators. They don't preserve func attributes.
-@track_analytics('API Hit', 'Update application', record_value_of_which_arg='application_id')
+@google_analytics_event('API Hit', 'Update application', record_value_of_which_arg='application_id')
 def update_application(application_id):
     # get the data from the request
     try:
@@ -189,7 +190,7 @@ def update_application(application_id):
 @blueprint.route("/applications/<application_id>", methods=["DELETE"])
 @api_key_required
 @swag(swag_summary='Delete an application <span class="red">[Authenticated, not public]</span>', swag_spec=ApplicationsCrudApi.delete_swag())  # must be applied after @api_key_(optional|required) decorators. They don't preserve func attributes.
-@track_analytics('API Hit', 'Delete application', record_value_of_which_arg='application_id')
+@google_analytics_event('API Hit', 'Delete application', record_value_of_which_arg='application_id')
 def delete_application(application_id):
     ApplicationsCrudApi.delete(application_id, current_user)
     return no_content()
@@ -201,7 +202,7 @@ def delete_application(application_id):
 @blueprint.route("/articles", methods=["POST"])
 @api_key_required
 @swag(swag_summary='Create an article <span class="red">[Authenticated, not public]</span>', swag_spec=ArticlesCrudApi.create_swag())  # must be applied after @api_key_(optional|required) decorators. They don't preserve func attributes.
-@track_analytics('API Hit', 'Create article')
+@google_analytics_event('API Hit', 'Create article')
 def create_article():
     # get the data from the request
     try:
@@ -219,7 +220,7 @@ def create_article():
 @blueprint.route("/articles/<article_id>", methods=["GET"])
 @api_key_optional
 @swag(swag_summary='Retrieve an article', swag_spec=ArticlesCrudApi.retrieve_swag())  # must be applied after @api_key_(optional|required) decorators. They don't preserve func attributes.
-@track_analytics('API Hit', 'Retrieve article', record_value_of_which_arg='article_id')
+@google_analytics_event('API Hit', 'Retrieve article', record_value_of_which_arg='article_id')
 def retrieve_article(article_id):
     a = ArticlesCrudApi.retrieve(article_id, current_user)
     return jsonify_models(a)
@@ -228,7 +229,7 @@ def retrieve_article(article_id):
 @blueprint.route("/articles/<article_id>", methods=["PUT"])
 @api_key_required
 @swag(swag_summary='Update an article <span class="red">[Authenticated, not public]</span>', swag_spec=ArticlesCrudApi.update_swag())  # must be applied after @api_key_(optional|required) decorators. They don't preserve func attributes.
-@track_analytics('API Hit', 'Update article', record_value_of_which_arg='article_id')
+@google_analytics_event('API Hit', 'Update article', record_value_of_which_arg='article_id')
 def update_article(article_id):
     # get the data from the request
     try:
@@ -246,7 +247,7 @@ def update_article(article_id):
 @blueprint.route("/articles/<article_id>", methods=["DELETE"])
 @api_key_required
 @swag(swag_summary='Delete an article <span class="red">[Authenticated, not public]</span>', swag_spec=ArticlesCrudApi.delete_swag())  # must be applied after @api_key_(optional|required) decorators. They don't preserve func attributes.
-@track_analytics('API Hit', 'Delete article', record_value_of_which_arg='article_id')
+@google_analytics_event('API Hit', 'Delete article', record_value_of_which_arg='article_id')
 def delete_article(article_id):
     ArticlesCrudApi.delete(article_id, current_user)
     return no_content()
@@ -258,7 +259,7 @@ def delete_article(article_id):
 @blueprint.route('/journals/<journal_id>', methods=['GET'])
 @api_key_optional
 @swag(swag_summary='Retrieve a journal by ID', swag_spec=JournalsCrudApi.retrieve_swag())  # must be applied after @api_key_(optional|required) decorators. They don't preserve func attributes.
-@track_analytics('API Hit', 'Retrieve journal', record_value_of_which_arg='journal_id')
+@google_analytics_event('API Hit', 'Retrieve journal', record_value_of_which_arg='journal_id')
 def retrieve_journal(journal_id):
     return jsonify_data_object(JournalsCrudApi.retrieve(journal_id, current_user))
 
@@ -269,7 +270,7 @@ def retrieve_journal(journal_id):
 @blueprint.route("/bulk/applications", methods=["POST"])
 @api_key_required
 @swag(swag_summary='Create applications in bulk <span class="red">[Authenticated, not public]</span>', swag_spec=ApplicationsBulkApi.create_swag())  # must be applied after @api_key_(optional|required) decorators. They don't preserve func attributes.
-@track_analytics('API Hit', 'Bulk application create')
+@google_analytics_event('API Hit', 'Bulk application create')
 def bulk_application_create():
     # get the data from the request
     try:
@@ -291,7 +292,7 @@ def bulk_application_create():
 @blueprint.route("/bulk/applications", methods=["DELETE"])
 @api_key_required
 @swag(swag_summary='Delete applications in bulk <span class="red">[Authenticated, not public]</span>', swag_spec=ApplicationsBulkApi.delete_swag())  # must be applied after @api_key_(optional|required) decorators. They don't preserve func attributes.
-@track_analytics('API Hit', 'Bulk application delete')
+@google_analytics_event('API Hit', 'Bulk application delete')
 def bulk_application_delete():
     # get the data from the request
     try:
@@ -309,7 +310,7 @@ def bulk_application_delete():
 @blueprint.route("/bulk/articles", methods=["POST"])
 @api_key_required
 @swag(swag_summary='Bulk article creation <span class="red">[Authenticated, not public]</span>', swag_spec=ArticlesBulkApi.create_swag())  # must be applied after @api_key_(optional|required) decorators. They don't preserve func attributes.
-@track_analytics('API Hit', 'Bulk article create')
+@google_analytics_event('API Hit', 'Bulk article create')
 def bulk_article_create():
     # get the data from the request
     try:
@@ -331,7 +332,7 @@ def bulk_article_create():
 @blueprint.route("/bulk/articles", methods=["DELETE"])
 @api_key_required
 @swag(swag_summary='Bulk article delete <span class="red">[Authenticated, not public]</span>', swag_spec=ArticlesBulkApi.delete_swag())  # must be applied after @api_key_(optional|required) decorators. They don't preserve func attributes.
-@track_analytics('API Hit', 'Bulk article create')
+@google_analytics_event('API Hit', 'Bulk article create')
 def bulk_article_delete():
     # get the data from the request
     try:
